Success and error messages for all entities, rework deleting category and tags
Create error message, refactor css and move admin css to path /resources/css/admin
Rework categories and tag deleting (if tag or category belong to product, they can`t be deleted)

+++ app/Http/Controllers/Category/DeleteController.php
@@ -4,12 +4,17 @@ namespace App\Http\Controllers\Category;
 
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Http\Request;
 
 class DeleteController extends Controller
 {
     public function __invoke(Category $category)
     {
+        $product = Product::where('category_id', $category->id)->get()->first();
+        if ($product) {
+            return back()->with(['error' => 'Категория id' . $category->id . ' привязана к товару, поэтому не может быть удалена']);
+        }
         $category->delete();
         return redirect()->route('category.index')->with(['success' => 'Категория id' . $category->id . ' успешно удалена']);
     }
